+++ lib/widgets/map/map_view.dart
@@ -5,7 +5,6 @@ import 'package:flutter_map/flutter_map.dart';
 import 'package:fml/log/manager.dart';
 import 'package:fml/observable/binding.dart';
 import 'package:flutter/material.dart';
-import 'package:fml/widgets/box/box_layout.dart';
 import 'package:fml/widgets/box/box_view.dart';
 import 'package:fml/widgets/map/map_model.dart';
 import 'package:fml/widgets/widget/widget_view_interface.dart';
@@ -199,8 +198,7 @@ class _MapViewState extends WidgetState<MapView> {
     Widget view = Stack(children: children);
 
     // create as Box
-    view = BoxView(widget.model,
-        children: [BoxLayout(model: widget.model, child: view)]);
+    view = BoxView(widget.model, children: [view]);
 
     return view;
   }
