+++ lib/widgets/modal/modal_model.dart
@@ -1,7 +1,6 @@
 // © COPYRIGHT 2022 APPDADDY SOFTWARE SOLUTIONS INC. ALL RIGHTS RESERVED.
 import 'package:flutter/material.dart';
 import 'package:fml/log/manager.dart';
-import 'package:fml/widgets/box/box_layout.dart';
 import 'package:fml/widgets/box/box_model.dart';
 import 'package:fml/widgets/modal/modal_manager_view.dart';
 import 'package:fml/widgets/widget/widget_model.dart';
@@ -191,13 +190,13 @@ class ModalModel extends BoxModel {
     List<Widget> views = [];
     for (var model in viewableChildren) {
       if (model is! ModalModel) {
+
         var view = model.getView();
 
         // wrap child in child data widget
         // this is done for us in "positioned" if the child happens
         // to be a positioned widget and the layout is "stack" (see positioned_view.dart)
         if (view != null) {
-          view = BoxLayout(model: model, child: view);
           views.add(view);
         }
       }
@@ -205,8 +204,7 @@ class ModalModel extends BoxModel {
 
     // add the static child
     if (child != null) {
-      var view = BoxLayout(model: this, child: child!);
-      views.add(view);
+      views.add(child!);
     }
 
     return views;
