+++ lib/widgets/scroller/scroller_model.dart
@@ -183,20 +183,34 @@ class ScrollerModel extends BoxModel implements IScrollable {
   /// scroll to specified item by id and value
   @override
   void scrollTo(String? id, String? value, {bool animate = false}) {
-    if (isNullOrEmpty(id)) return;
+
+    if (id == null) return;
+
+    // get the view
+    ScrollerViewState? view = findListenerOfExactType(ScrollerViewState);
+
+    // scroll to top
+    if (id.trim().toLowerCase() == 'top' && isNullOrEmpty(value)) {
+      view?.scrollTo(0, animate: false);
+      return;
+    }
+
+    // scroll to bottom
+    if (id.trim().toLowerCase() == 'bottom' && isNullOrEmpty(value)) {
+      view?.scrollTo(double.maxFinite, animate: false);
+      return;
+    }
+
+    // scroll to specific pixel position
+    if (isNumeric(id) && isNullOrEmpty(value)) {
+      view?.scrollTo(toDouble(id), animate: false);
+    }
 
     // find the first child with the specified
     // id and matching value
-    BuildContext? context;
     var child = descendants?.toList().firstWhereOrNull((child) => child.id == id && child.value == (value ?? child.value));
     if (child != null) {
-      context = child.context;
-    }
-
-    // context defined?
-    if (context != null) {
-      ScrollerViewState? view = findListenerOfExactType(ScrollerViewState);
-      view?.scrollTo(context, animate: animate);
+      view?.scrollToContext(child.context, animate: animate);
     }
   }
