+++ lib/widgets/treeview/tree_view.dart
@@ -2,7 +2,6 @@
 import 'package:flutter/material.dart';
 import 'package:fml/event/event.dart';
 import 'package:fml/event/manager.dart';
-import 'package:fml/widgets/box/box_layout.dart';
 import 'package:fml/widgets/box/box_view.dart';
 import 'package:fml/widgets/treeview/tree_model.dart';
 import 'package:fml/widgets/treeview/node/tree_node_view.dart';
@@ -65,8 +64,7 @@ class _TreeViewState extends WidgetState<TreeView> {
                 TreeNodeView(widget.model.nodes[index])));
 
     // create as Box
-    view = BoxView(widget.model,
-        children: [BoxLayout(model: widget.model, child: view)]);
+    view = BoxView(widget.model, children: [view]);
 
     return view;
   }
